feat(ckwu): add 1-shot implementation (sample from existing demos)

+++ experiment.py
@@ -1,7 +1,9 @@
+import os
 import math
 import json
 import yaml
 import pandas as pd
+import random
 from typing import Union
 from pathlib import Path
 from argparse import ArgumentParser, Namespace
@@ -79,7 +81,8 @@ class Experiment(object):
         task_continue_from: str = None,
         sample_start_from: int = 0,
         label_type: str = None,
-        lacked_cases_path: Path = None
+        lacked_cases_path: Path = None,
+        existing_demos_path: Path = None
     ) -> None:
         # handle passed arguments
         if label_type and (label_type not in set(TaskGenerator.task2label_type.values())):
@@ -172,38 +175,55 @@ class Experiment(object):
                     )
                 # augment prompt with pseudo-demos if "self-icl"
                 if self._config.exemplars_mode == "self-icl":
-                    # generate pseudo-demos
-                    # inputs
-                    demo_prompt = prompt.gen_demo_inputs(diversity=self._config.diverse_exemplars)
-                    demo_prompt, demo_inputs = self._model.complete(demo_prompt, label_set=None, temperature=self._config.demo_temperature)
-                    full_demo_inputs = demo_prompt + demo_inputs
-                    (task_log_path / "demo-inputs" / f"{i}.txt").write_text(full_demo_inputs)
-                    # parse demo inputs to separate instances
-                    try:
-                        sep_demo_inputs = self._prompt_parser.split_demo_inputs(full_demo_inputs)
-                    except ValueError:
-                        print(Fore.RED + f"Task {task_name} sample #{i} failed: failed to parse demo inputs" + Style.RESET_ALL)
-                        failed_cases.append([task_name, i])
-                        continue
-                    # labels
-                    shots = []
-                    if self._config.inference_mode == "stream":
-                        for j, sep_demo_input in enumerate(sep_demo_inputs):
-                            sep_demo_prompt = StreamPrompt(
-                                task_desc=task.task_desc,
-                                inputs=sep_demo_input,
-                                num_demos=0, # NOTE
-                                shots=[]
-                            ).gen_prediction(cot=self._config.use_cot)
-                            print(f"Predicting demo #{j} (cot: {self.cot_check if self._config.use_cot else self.cot_cross}) -> ", end='')
-                            sep_demo_prompt, sep_demo_label = self._model.complete(sep_demo_prompt, label_set, temperature=self._config.temperature)
-                            if sep_demo_prompt[-1] == '(':
-                                sep_demo_label = '(' + sep_demo_label
-                            shot = Shot(_input=sep_demo_input, _label=sep_demo_label.strip())
+                    if existing_demos_path is None:
+                        # generate pseudo-demos
+                        # inputs
+                        demo_prompt = prompt.gen_demo_inputs(diversity=self._config.diverse_exemplars)
+                        demo_prompt, demo_inputs = self._model.complete(demo_prompt, label_set=None, temperature=self._config.demo_temperature)
+                        full_demo_inputs = demo_prompt + demo_inputs
+                        (task_log_path / "demo-inputs" / f"{i}.txt").write_text(full_demo_inputs)
+                        # parse demo inputs to separate instances
+                        try:
+                            sep_demo_inputs = self._prompt_parser.split_demo_inputs(full_demo_inputs)
+                        except ValueError:
+                            print(Fore.RED + f"Task {task_name} sample #{i} failed: failed to parse demo inputs" + Style.RESET_ALL)
+                            failed_cases.append([task_name, i])
+                            continue
+                    else:
+                        num_existing_demos = len(os.listdir(existing_demos_path / task_name / "demo-labels")) // task.sample_size
+                        if self._config.num_demos == num_existing_demos:
+                            js = list(range(num_existing_demos))
+                        else:
+                            js = random.sample(range(num_existing_demos), self._config.num_demos)
+                        for j in js:
+                            demo_input_label = (existing_demos_path / task_name / "demo-labels" / f"{i}-{j}.txt").read_text()
+                            q_start = demo_input_label.index("Q:")
+                            a_start = demo_input_label.index("A:")
+                            demo_input = demo_input_label[q_start+len("Q:"):a_start].strip()
+                            demo_label = demo_input_label[a_start+len("A:"):].strip()
+                            print(f"Adding existing demo #{j} -> {demo_label}")
+                            shot = Shot(_input=demo_input, _label=demo_label)
                             shots.append(shot)
-                            # logging
-                            print(sep_demo_label)
-                            (task_log_path / "demo-labels" / f"{i}-{j}.txt").write_text(str(shot))
+                    # labels
+                    if self._config.inference_mode == "stream":
+                        if existing_demos_path is None:
+                            shots = []
+                            for j, sep_demo_input in enumerate(sep_demo_inputs):
+                                sep_demo_prompt = StreamPrompt(
+                                    task_desc=task.task_desc,
+                                    inputs=sep_demo_input,
+                                    num_demos=0, # NOTE
+                                    shots=[]
+                                ).gen_prediction(cot=self._config.use_cot)
+                                print(f"Predicting demo #{j} (cot: {self.cot_check if self._config.use_cot else self.cot_cross}) -> ", end='')
+                                sep_demo_prompt, sep_demo_label = self._model.complete(sep_demo_prompt, label_set, temperature=self._config.temperature)
+                                if sep_demo_prompt[-1] == '(':
+                                    sep_demo_label = '(' + sep_demo_label
+                                shot = Shot(_input=sep_demo_input, _label=sep_demo_label.strip())
+                                shots.append(shot)
+                                # logging
+                                print(sep_demo_label)
+                                (task_log_path / "demo-labels" / f"{i}-{j}.txt").write_text(str(shot))
                         # update prompt to augmented prompt
                         prompt = StreamPrompt(
                             task_desc=task.task_desc,
@@ -217,9 +237,20 @@ class Experiment(object):
                             inputs=sep_demo_inputs,
                             num_demos=0, # NOTE
                             shots=shots
-                        )        
-                        raise NotImplementedError
-                        # update prompt to augmented prompt
+                        ).gen_prediction(add_parenthesis=add_parenthesis)
+                        sep_demo_prompt, sep_demo_label = self._model.complete(sep_demo_prompt, label_set, temperature=self._config.temperature)
+                        # update prompt to augmented prompt (shots == full_demo)
+                        shots = sep_demo_prompt + sep_demo_label
+                        shots = shots[shots.index("Q1:"):] # remove task description
+                        print(sep_demo_label)
+                        (task_log_path / "demo-labels" / f"{i}.txt").write_text(str(shots))
+
+                        prompt = BatchPrompt(
+                            task_desc=task.task_desc, # shots already contain task description in current implementation
+                            inputs=task_inputs,
+                            num_demos=self._config.num_demos,
+                            shots=shots
+                        )
                     
                 # run inference
                 print(f"Predicting sample #{i} (cot: {self.cot_check if self._config.use_cot else self.cot_cross}) ->", end='')
@@ -328,6 +359,7 @@ def parse_args() -> Namespace:
     parser.add_argument("--label_type", type=str, default=None)
     parser.add_argument("--eval", action="store_true")
     parser.add_argument("--weighted_acc", action="store_true")
+    parser.add_argument("--existing_demos_path", type=Path, default=None)
     return parser.parse_args()
 
 if __name__ == "__main__":
@@ -345,5 +377,6 @@ if __name__ == "__main__":
             task_continue_from=args.task_continue_from,
             sample_start_from=args.sample_start_from,
             label_type=args.label_type,
-            lacked_cases_path=args.lacked_cases_path
+            lacked_cases_path=args.lacked_cases_path,
+            existing_demos_path=args.existing_demos_path
         )
